requirements.txt
numpy==1.21.6
Pillow==9.5.0
Pillow==10.0.0
scikit_image==0.19.3
ultralytics==8.0.145
